refactor(ui): extract inventory table helpers into focused component #164
Extract StatusBadge, HealthBar, and SortHeader sub-components from
inventory.tsx (458 lines) into inventory/device-table-helpers.tsx,
bringing the main file to 367 lines (under 400-line threshold).
